requirements.txt
Flask>=2.0
PyMuPDF>=1.18
python-docx>=0.8.11
sentence_transformers